archive/core/nia-chat-agent/requirements.txt
groq
requests
psutil
termcolor
streamlit
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib
python-dotenv
pydantic
twilio
fastapi
pymongo

